app: Playlist Burning Happens All At Once
Instead of incrementally adding songs to a tempdir that we transcode +
downsample to while the user selects song, we do it all at once when we
burn the playlist since it will already take a while to burn the CD. It
also enables the use of building the playlist without the computer
having access to the files until playlist burn time.

Instead of having playlist burn log directly to stdout/stderr, we now
have a channel we write log messages to. This will be used to separate
the implementation from the view for later use in the TUI.

Add ability to add songs directly to the playlist without querying the
database.